+++ js/executable_scripts/point_of_interest.js
