+++ js/popup/asset_downloader.js
@@ -1,0 +1,5 @@
+function transfer() {
+    chrome.tabs.executeScript({ file: "scripts/jquery-3.2.1.min.js" }, function () {
+      chrome.tabs.executeScript({ file: "functions/transfer.js" });
+    });
+  }
