+++ js/popup/auto_alt.js
@@ -1,0 +1,5 @@
+function alt() {
+  chrome.tabs.executeScript({ file: "scripts/jquery-3.2.1.min.js" }, function () {
+    chrome.tabs.executeScript({ file: "functions/auto_alt.js" });
+  });
+}
