+++ js/popup/auto_redirects.js
@@ -1,0 +1,5 @@
+function redirects() {
+  chrome.tabs.executeScript({ file: "scripts/jquery-3.2.1.min.js" }, function () {
+    chrome.tabs.executeScript({ file: "functions/redirects.js" });
+  });
+}
