+++ js/popup/google_cid.js
@@ -1,0 +1,5 @@
+function googleCID() {
+    chrome.tabs.executeScript({ file: "scripts/jquery-3.2.1.min.js" }, function () {
+      chrome.tabs.executeScript({ file: "functions/googleCID.js" });
+    });
+  }
