+++ js/popup/hub_search.js
@@ -1,0 +1,3 @@
+function hubSearch() {
+    chrome.tabs.executeScript({ file: "functions/hub-Search.js" });
+  }
