+++ js/popup/open_sidekiq.js
@@ -1,0 +1,9 @@
+function sidekiq() {
+    chrome.tabs.query({ active: true, currentWindow: true }, function (tabs) {
+      var currentURL = tabs[0].url;
+      // https://stackoverflow.com/questions/3689423/google-chrome-plugin-how-to-get-domain-from-url-tab-url
+      var domain = currentURL.match(/^[\w-]+:\/{2,}\[?[\w\.:-]+\]?(?::[0-9]*)?/)[0];
+      var url = domain + '/sidekiq/busy?poll=true';
+      chrome.tabs.create({ url: url });
+    });
+  }
