+++ js/popup/seo_updater.js
@@ -1,0 +1,5 @@
+function seo() {
+  chrome.tabs.executeScript({ file: "scripts/jquery-3.2.1.min.js" }, function () {
+    chrome.tabs.executeScript({ file: "functions/seo_updater.js" });
+  });
+}
