+++ popup.html
@@ -133,7 +133,7 @@
   <script type="text/javascript" src="/bootstrap/popper.js"></script>
   <script type="text/javascript" src="/bootstrap/bootstrap.min.js"></script>
   <script src="/scripts/papaparse.min.js"></script>
-  <script type="text/javascript" src="popup.js"></script>
+  <script type="module" src="popup.js"></script>
 </body>
 
 </html>
